+++ NewHaven/CMakeLists.txt
@@ -13,6 +13,6 @@ if(UNIX)
     #Execute Unix only command
     endif(UNIX)
 
-add_executable(NewHaven main.cpp "source/GBMap.cpp")
+add_executable(NewHaven GBMapDriver.cpp "source/GBMap.cpp")
 
 target_link_libraries (NewHaven ${Boost_LIBRARIES})
